port balance and all money operations from double to BigDecimal

+++ src/main/java/com/themescaline/moneytransfer/dao/AccountDAO.java
@@ -2,6 +2,7 @@ package com.themescaline.moneytransfer.dao;
 
 import com.themescaline.moneytransfer.model.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -25,9 +26,9 @@ public interface AccountDAO {
 
     void delete(long accountId);
 
-    void transfer(long fromAccountId, long toAccountId, double amount);
+    void transfer(long fromAccountId, long toAccountId, BigDecimal amount);
 
-    void withdraw(long accountId, double amount);
+    void withdraw(long accountId, BigDecimal amount);
 
-    void deposit(long accountId, double amount);
+    void deposit(long accountId, BigDecimal amount);
 }
